update hyperloop module to 1.1.4
