+++ modules/android/hyperloop/1.1.4/manifest
@@ -1,0 +1,18 @@
+#
+# this is your module manifest and used by Titanium
+# during compilation, packaging, distribution, etc.
+#
+version: 1.1.4
+apiversion: 2
+architectures: armeabi armeabi-v7a x86
+description: hyperloop-android
+author: Appcelerator
+license: Appcelerator Commercial License
+copyright: Copyright (c) 2016 Appcelerator, Inc.
+
+# these should not be edited
+name: hyperloop-android
+moduleid: hyperloop
+guid: bdaca69f-b316-4ce6-9065-7a61e1dafa39
+platform: android
+minsdk: 5.2.1
