--- modules/iphone/hyperloop/1.1.2/libhyperloop.a
@@ -1,1 +1,0 @@
-1
