--- modules/iphone/hyperloop/1.1.2/manifest
@@ -1,18 +1,0 @@
-#
-# this is your module manifest and used by Titanium
-# during compilation, packaging, distribution, etc.
-#
-version: 1.1.2
-apiversion: 2
-architectures: armv7 arm64 i386 x86_64
-description: hyperloop
-author: Appcelerator
-license: Appcelerator Commercial License
-copyright: Copyright (c) 2015 by Appcelerator, Inc.
-
-# these should not be edited
-name: hyperloop
-moduleid: hyperloop
-guid: bdaca69f-b316-4ce6-9065-7a61e1dafa39
-platform: iphone
-minsdk: 5.4.0
