--- modules/iphone/hyperloop/1.1.2/module.xcconfig
@@ -1,27 +1,0 @@
-//
-// PLACE ANY BUILD DEFINITIONS IN THIS FILE AND THEY WILL BE
-// PICKED UP DURING THE APP BUILD FOR YOUR MODULE
-//
-// see the following webpage for instructions on the settings
-// for this file:
-// http://developer.apple.com/mac/library/documentation/DeveloperTools/Conceptual/XcodeBuildSystem/400-Build_Configurations/build_configs.html
-//
-
-//
-// How to add a Framework (example)
-//
-// OTHER_LDFLAGS=$(inherited) -framework Foo
-//
-// Adding a framework for a specific version(s) of iPhone:
-//
-// OTHER_LDFLAGS[sdk=iphoneos4*]=$(inherited) -framework Foo
-// OTHER_LDFLAGS[sdk=iphonesimulator4*]=$(inherited) -framework Foo
-//
-//
-// How to add a compiler define:
-//
-// OTHER_CFLAGS=$(inherited) -DFOO=1
-//
-//
-// IMPORTANT NOTE: always use $(inherited) in your overrides
-//
