+++ plugins/hyperloop/hooks/ios/hyperloop.js
@@ -55,7 +55,7 @@ function HyperloopiOSBuilder(logger, config, cli, appc, hyperloopConfig, builder
 	this.resourcesDir = path.join(builder.projectDir, 'Resources');
 	this.hyperloopBuildDir = path.join(builder.projectDir, 'build', 'hyperloop', 'ios');
 	this.hyperloopJSDir = path.join(this.hyperloopBuildDir, 'js');
-	this.hyperloopResourcesDir = path.join(this.resourcesDir, 'hyperloop');
+	this.hyperloopResourcesDir = path.join(this.resourcesDir, 'iphone', 'hyperloop');
 
 	this.forceMetabase = false;
 	this.forceStubGeneration = false;
@@ -496,6 +496,7 @@ HyperloopiOSBuilder.prototype.generateSourceFiles = function generateSourceFiles
 				this.frameworks.$metadata.sdkType,
 				this.frameworks.$metadata.sdkPath,
 				this.frameworks.$metadata.minVersion,
+				this.builder.xcodeTargetOS,
 				this.metabase,
 				entry.framework,
 				entry.source,
@@ -526,9 +527,13 @@ HyperloopiOSBuilder.prototype.generateSourceFiles = function generateSourceFiles
  * Generates the symbol reference based on the references from the metabase's parser state.
  */
 HyperloopiOSBuilder.prototype.generateSymbolReference = function generateSymbolReference() {
-	var symbolRefFile = path.join(this.hyperloopBuildDir, 'symbol_references.json');
-	var json = JSON.stringify(this.parserState.getReferences(), null, 2);
 
+	if (!this.parserState) {
+		this.logger.info('Skipping ' + HL + ' generating of symbol references. Empty AST. ');
+		return;
+	}	
+	var symbolRefFile = path.join(this.hyperloopBuildDir, 'symbol_references.json'),
+		json = JSON.stringify(this.parserState.getReferences(), null, 2);
 	if (!fs.existsSync(symbolRefFile) || fs.readFileSync(symbolRefFile).toString() !== json) {
 		this.forceStubGeneration = true;
 		this.logger.trace('Forcing regeneration of wrappers');
@@ -550,6 +555,11 @@ HyperloopiOSBuilder.prototype.compileResources = function compileResources(callb
  * Generates stubs from the metabase.
  */
 HyperloopiOSBuilder.prototype.generateStubs = function generateStubs(callback) {
+
+	if (!this.parserState) {
+		this.logger.info('Skipping ' + HL + ' stub generation. Empty AST.');
+		return callback();
+	}
 	if (!this.forceStubGeneration) {
 		this.logger.debug('Skipping stub generation');
 		return callback();
@@ -921,7 +931,8 @@ HyperloopiOSBuilder.prototype.hookUpdateXcodeProject = function hookUpdateXcodeP
 			xobjs.PBXBuildFile[buildFileUuid] = {
 				isa: 'PBXBuildFile',
 				fileRef: fileRefUuid,
-				fileRef_comment: name
+				fileRef_comment: name,
+				settings: {COMPILER_FLAGS : '"-fobjc-arc"' }
 			};
 			xobjs.PBXBuildFile[buildFileUuid + '_comment'] = name + ' in Sources';
