+++ plugins/hyperloop/package.json
@@ -1,6 +1,6 @@
 {
   "name": "hyperloop",
-  "version": "1.1.2",
+  "version": "1.1.4",
   "description": "Provides direct native API access from Titanium",
   "keywords": [
     "appcelerator",
