+++ tiapp.xml
@@ -62,7 +62,7 @@
 		<theme>default</theme>
 	</mobileweb>
 	<modules>
-		<module version="1.1.2">hyperloop</module>
+		<module version="1.1.4">hyperloop</module>
 	</modules>
 	<deployment-targets>
 		<target device="android">true</target>
